+++ controller/konnect/reconciler_generic_test.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/kong/gateway-operator/modules/manager/scheme"
 
+	configurationv1 "github.com/kong/kubernetes-configuration/api/configuration/v1"
 	configurationv1alpha1 "github.com/kong/kubernetes-configuration/api/configuration/v1alpha1"
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
@@ -17,14 +18,11 @@ import (
 func TestNewKonnectEntityReconciler(t *testing.T) {
 	testNewKonnectEntityReconciler(t, konnectv1alpha1.KonnectControlPlane{})
 	testNewKonnectEntityReconciler(t, configurationv1alpha1.KongService{})
+	testNewKonnectEntityReconciler(t, configurationv1.KongConsumer{})
 
 	// TODO(pmalek): add support for KongRoute
 	// https://github.com/Kong/gateway-operator/issues/435
 	// testNewKonnectEntityReconciler(t, configurationv1alpha1.KongRoute{})
-
-	// TODO(pmalek): add support for KongConsumer
-	// https://github.com/Kong/gateway-operator/issues/436
-	// testNewKonnectEntityReconciler(t, configurationv1.KongConsumer{})
 
 	// TODO: GetConditions() and SetConditions() is missing from KongConsumerGroup.
 	// testNewKonnectEntityReconciler(t, configurationv1beta1.KongConsumerGroup{})